fix: read builder config
