+++ .changeset/hungry-balloons-hear.md
@@ -1,0 +1,6 @@
+---
+'@modern-js/storybook-builder': patch
+---
+
+feat(storybook): add adapter for storybook addons, read builderConfig
+feat(storybook): 为 storybook addons 添加一些适配逻辑，读取builderConfig
